sync

+++ codec/src/main/java/io/netty/handler/codec/compression/Snappy.java
@@ -607,7 +607,7 @@ public final class Snappy {
         Crc32c crc32 = new Crc32c();
         try {
             crc32.update(data, offset, length);
-            return maskChecksum((int) crc32.getValue());
+            return maskChecksum(crc32.getValue());
         } finally {
             crc32.reset();
         }
@@ -655,7 +655,7 @@ public final class Snappy {
      * @param checksum The actual checksum of the data
      * @return The masked checksum
      */
-    static int maskChecksum(int checksum) {
-        return (checksum >> 15 | checksum << 17) + 0xa282ead8;
+    static int maskChecksum(long checksum) {
+        return (int) ((checksum >> 15 | checksum << 17) + 0xa282ead8);
     }
 }
